// src/core/types.ts
export type LocaleTypes = 'en' | 'ar'

export type ParamsTypes = {
    params: { slug: IdTypes }
}

export type InfoType = {
    id: IdTypes
    name: string
}

export type PersonTypes = InfoType & {
    cover: string
}

export type CardNameTypes = 'album' 
| 'artist' 
| 'avatar' 
| 'playlist' 
| 'genre' 
| 'event' 
| 'radio' 
| 'track'


/**
 * 
 * Theme types
 */
export type ThemeModeTypes = 'light' | 'dark' | 'system'
export type ComponentNameTypes = 'header' | 'sidebar' | 'player'
export type ComponentSkinTypes = 'red' 
| 'green' 
| 'blue' 
| 'orange' 
| 'yellow' 
| 'purple' 
| 'indigo'
| 'pink' 
| 'violet' 
| 'magenta'
// ...
export type NavLinkTypes = {
    name: string
    href: string
    icon?: any
    fragment?: string
}

export type NavHeadTypes = {
    title: string
}

export type NavbarTypes = NavLinkTypes | NavHeadTypes